requirements.txt
Django==4.2.6
djangorestframework==3.14.0
psycopg2==2.9.9
bcrypt==4.0.1
PyJWT==2.8.0